Swapchain image count of 4 on some devices
On some devices, like intel integrated graphics cards, the minimum image count is 3. When adding 1 to this, there are 4 images which causes issues because the number of sync objects is hard coded as 2. Another fix for this would be to create imageCount - 1 sync objects, but that would result in quadruple buffering on the devices with 3 images.